+++ encode_jpeg.h
@@ -17,7 +17,7 @@
 #include "encode_jpeg.h"
 #include "fill_bits.h"
 
-	void encode_jpeg_init(struct fb_var_screeninfo sc, int fb_effective_bytes_per_pixel);
+	void encode_jpeg_init(struct fb_var_screeninfo sc, uint8_t fb_effective_bytes_per_pixel);
 	uint8_t *encode_jpeg(uint8_t fb_effective_bytes_per_pixel, void *fb_buf_1dim, uint32_t *imagesize);
 	void encode_jpeg_finalize();
