+++ read_fb.h
@@ -9,7 +9,7 @@
 #include <sys/types.h>
 #include <linux/fb.h>
 
-	void read_fb_init(char *devname, struct fb_var_screeninfo *sc, uint32_t *effective_bytes_per_pixel, uint64_t *size);
+	void read_fb_init(char *devname, struct fb_var_screeninfo *sc, uint8_t *effective_bytes_per_pixel, uint64_t *size);
 	void read_fb(char *buf, uint32_t size);
 	void read_fb_finalize();
